Finally fixed the UI (for now)
When the user gets into a vinicity (set to 10 feet for now), then the screen turns green, as a clear visual indicator.

The circle UI broke, but I'll improve it later. It's not great at the moment so I need to come up with a better design anyways

Breakdown: Instead of trying to set distanceInFeet directly inside the view, the updateDistance() function is called whenever latestLocation changes. The updateColors(for:) function is responsible for changing the colors according to the distance.